Refactor dir
